LayoutTests/svg/custom/poly-parsing-error-expected.txt

CONSOLE MESSAGE: Error: Problem parsing points="80,200 80,300 150,250 80,200 250"
CONSOLE MESSAGE: Error: Problem parsing points="180,200 180,300 250,250 180,200 250"
CONSOLE MESSAGE: Error: Problem parsing points="80,60 80,160 150,110 80"
CONSOLE MESSAGE: Error: Problem parsing points="180,60 180,160 250,110 180"
Tests whether polygons clear items on parsing error per SVG2 spec.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS poly1.points.numberOfItems is 0
PASS poly2.points.numberOfItems is 4
PASS poly3.points.numberOfItems is 4
PASS poly4.points.numberOfItems is 0
PASS poly5.points.numberOfItems is 0
PASS poly6.points.numberOfItems is 3
PASS poly7.points.numberOfItems is 3
PASS poly8.points.numberOfItems is 0
PASS successfullyParsed is true

TEST COMPLETE
